make common component
